Changed tests2 to be the new tests
